requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy
io
zipfile